DEV: adding util to calculate wpi from args (temp, wind, connnndition)
